Rewrite chart to use config as-is, keys in Secret
Instead of exposing many values and generating the config from them in
the container, we expose the entire config and put it in a ConfigMap.

We still expose a extraConfig mechanism to override specific values from
Secrets or ConfigMaps.

In addition, we store the secret keys in a Secret, and use a Job to
generate them when the Chart is installed.